+++ packages/expo-doctor/src/checks/PackageJsonCheck.ts
@@ -42,6 +42,27 @@ export class PackageJsonCheck implements DoctorCheck {
       );
     }
 
+    // ** check if the package.json contains react 19 and if an override is set for it. If an override is not set this can cause a runtime issue:
+    // "Warning: Error: A React Element from an older version of React was rendered. This is not supported"
+    //  **
+
+    const reactVersion = pkg.dependencies?.react;
+    const reactOverride = pkg.overrides?.react;
+
+    if (
+      reactVersion &&
+      reactVersion.includes('19.') &&
+      (!reactOverride || !reactOverride.includes('19.'))
+    ) {
+      issues.push(
+        'React 19 is installed but no override is set for it in the package.json. To prevent runtime issues, please add the following override to your package.json:\n' +
+          `\n"overrides": {\n` +
+          `  "react": "${reactVersion}"\n` +
+          `}\n` +
+          `\nThis will ensure that react-native will always use the correct version of React.`
+      );
+    }
+
     return {
       isSuccessful: issues.length === 0,
       issues,